+++ client/videowatch/share.ts
@@ -1,0 +1,11 @@
+async function shareChatUrl ({ peertubeHelpers }: RegisterOptions): Promise<void> {
+  peertubeHelpers.showModal({
+    title: 'TODO',
+    content: '<p>TODO</p>',
+    close: true
+  })
+}
+
+export {
+  shareChatUrl
+}